+++ CHANGELOG.md
@@ -37,7 +37,8 @@ rewritten to their query representation
 - Metas attached to `INTERVAL` `PType`s if an `INTERVAL` has an unspecified precision or fractional precision
 - `EXTRACT` function support on `INTERVAL` datum
 - `ABS` function support on `INTERVAL` values
-- support for `OVERLAPS` predicate
+- Support for `OVERLAPS` predicate
+- Add cli option -e to load PartiQL literal data as default catalog data
 
 ### Changed